+++ plugins/tasks/slack/src/main/java/com/walmartlabs/concord/plugins/slack/SlackConfiguration.java
@@ -1,0 +1,80 @@
+package com.walmartlabs.concord.plugins.slack;
+
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2018 Wal-Mart Store, Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+public class SlackConfiguration {
+
+    public static final int DEFAULT_CONNECT_TIMEOUT = 30_000;
+    public static final int DEFAULT_SO_TIMEOUT = 30_000;
+    public static final int DEFAULT_RETRY_COUNT = 5;
+
+    private final String authToken;
+    private String proxyAddress;
+    private Integer proxyPort;
+    private int connectTimeout = DEFAULT_CONNECT_TIMEOUT;
+    private int soTimeout = DEFAULT_SO_TIMEOUT;
+    private int retryCount = DEFAULT_RETRY_COUNT;
+
+    public SlackConfiguration(String authToken) {
+        this.authToken = authToken;
+    }
+
+    public void setProxy(String proxyAddress, Integer proxyPort) {
+        this.proxyAddress = proxyAddress;
+        this.proxyPort = proxyPort;
+    }
+
+    public void setConnectTimeout(int connectTimeout) {
+        this.connectTimeout = connectTimeout;
+    }
+
+    public void setSoTimeout(int soTimeout) {
+        this.soTimeout = soTimeout;
+    }
+
+    public void setRetryCount(int retryCount) {
+        this.retryCount = retryCount;
+    }
+
+    public String getAuthToken() {
+        return authToken;
+    }
+
+    public String getProxyAddress() {
+        return proxyAddress;
+    }
+
+    public Integer getProxyPort() {
+        return proxyPort;
+    }
+
+    public int getConnectTimeout() {
+        return connectTimeout;
+    }
+
+    public int getSoTimeout() {
+        return soTimeout;
+    }
+
+    public int getRetryCount() {
+        return retryCount;
+    }
+}