+++ sdk/src/main/java/com/walmartlabs/concord/sdk/RpcClient.java
@@ -27,6 +27,4 @@ public interface RpcClient {
     SecretStoreService getSecretStoreService();
 
     EventService getEventService();
-
-    SlackService getSlackService();
 }